This commit addresses several outstanding issues with the harness.
1. It fixes the parsing bug in `janice_enroll_detections` from issue #32.
2. IJB-C compliant headers are parsed properly. This should fix issue #20.
3. The harness programs output CSVs consistent with the recent data calls.
The input requirements for `janice_search` and `janice_verify` are changed
to require only template IDs. This is consistent with the new output of
`janice_enroll_media` and `janice_enroll_detections`. Some expected fields
in the output can't be reliably filled in by the harness. For example,
timing information is not accurate because of the use of batch functions.
These fields are filled with `-1`. This should fix issue #35.